+++ core/pioreactor/calibrations/registry.py
@@ -20,6 +20,7 @@ class CalibrationProtocol(Generic[Device]):
     title: ClassVar[str] = ""
     description: ClassVar[str] = ""
     requirements: ClassVar[tuple[str, ...]] = ()
+    step_registry: ClassVar[dict[str, type[Any]]]
 
     def __init_subclass__(cls, **kwargs):
         super().__init_subclass__(**kwargs)